# sample/service-user-center/docs/api/single.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/permission": {
      "post": {
        "operationId": "PermissionHttpController_insert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionInsertReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionInsertResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      },
      "get": {
        "operationId": "PermissionHttpController_selectByPage",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sysId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "extInfo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSelectByPageResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      }
    },
    "/api/permission/{id}": {
      "get": {
        "operationId": "PermissionHttpController_selectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSelectByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      },
      "patch": {
        "operationId": "PermissionHttpController_updateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionUpdateReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      },
      "delete": {
        "operationId": "PermissionHttpController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      }
    },
    "/api/permission/m": {
      "post": {
        "operationId": "PermissionHttpController_insertMulti",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionInsertMultiResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      }
    },
    "/api/permission/m/{ids}": {
      "get": {
        "operationId": "PermissionHttpController_selectByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSelectByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      },
      "patch": {
        "operationId": "PermissionHttpController_updateByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionUpdateReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      },
      "delete": {
        "operationId": "PermissionHttpController_deleteByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Permission - 权限信息"
        ]
      }
    },
...
  },
  "info": {
    "title": "SAMPLE SERVICE USER CENTER",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "PermissionInsertReqDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sysId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "extInfo": {
            "type": "string"
          }
        }
      },
      "PermissionInsertResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sysId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "extInfo": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ]
      },
      "PermissionInsertResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/PermissionInsertResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "PermissionSelectResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sysId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "extInfo": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ]
      },
      "PageResDto": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "number"
          },
          "pageNum": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "pageSize",
          "pageNum",
          "total"
        ]
      },
      "PermissionSelectByPageResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionSelectResDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "page",
          "message"
        ]
      },
      "PermissionSelectByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PermissionSelectResDto"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "PermissionUpdateReqDto": {
        "type": "object",
        "properties": {
          "sysId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "extInfo": {
            "type": "string"
          }
        }
      },
      "UpdateEffectInfo": {
        "type": "object",
        "properties": {
          "effectNum": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "effectNum"
        ]
      },
      "UpdateByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "DeleteEffectInfo": {
        "type": "object",
        "properties": {
          "effectNum": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "effectNum"
        ]
      },
      "DeleteByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "PermissionInsertMultiResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionInsertResDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "PermissionSelectByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionSelectResDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "UpdateByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "DeleteByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
...
    }
  }
}